+++ CHANGELOG.md
@@ -4,6 +4,7 @@ changes to any options/interfaces the checker exposes for developers.
 
 # 17.3.0
 26 March 2017
+  - Allow “color” attribute with link[rel="mask-icon"]
   - Allow `meta[name]` to have `itemref`/`itemscope`/`itemtype`/`itemid`
   - Allow `allow-top-navigation-by-user-activation` in `iframe[sandbox]`
   - Stop hiding “sectioning roots” headings in “Heading-level outline”